# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Project Management Platform API
  version: 0.2.0
  description: |
    Refactor contract-first API baseline.
    Due date semantic migration: `dueAt` is canonical; `dueDate` accepted for backward compatibility in write payloads and exposed only in legacy `/tasks`.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}/issues:
    get:
      summary: List issues by project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: statusId
          schema: { type: string }
        - in: query
          name: assigneeId
          schema: { type: string }
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [number, updatedAt]
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Paged issue list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Create issue
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
  /issues/{issueId}/status:
    patch:
      summary: Transition issue status
      parameters:
        - $ref: '#/components/parameters/IssueId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [statusId]
              properties:
                statusId:
                  type: string
      responses:
        '200':
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '422':
          description: Unknown or invalid transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /workflows/statuses:
    get:
      summary: List workflow statuses
      responses:
        '200':
          description: Workflow statuses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkflowStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /projects/{projectId}/board:
    get:
      summary: Get project board columns
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Board columns
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BoardColumn'
        '403':
          $ref: '#/components/responses/Forbidden'
  /issues/{issueId}:
    get:
      summary: Get issue detail
      parameters:
        - $ref: '#/components/parameters/IssueId'
      responses:
        '200':
          description: Issue detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update issue fields
      parameters:
        - $ref: '#/components/parameters/IssueId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssueRequest'
      responses:
        '200':
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /issues/{issueId}/assignee:
    patch:
      summary: Assign an issue
      parameters:
        - $ref: '#/components/parameters/IssueId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assigneeId]
              properties:
                assigneeId: { type: [string, 'null'] }
      responses:
        '200':
          description: Assigned issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /issues/{issueId}/comments:
    get:
      summary: List issue comments
      parameters:
        - $ref: '#/components/parameters/IssueId'
      responses:
        '200':
          description: Issue comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Create an issue comment
      parameters:
        - $ref: '#/components/parameters/IssueId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        '201':
          description: Created comment
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Comment'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /activity-logs:
    get:
      summary: List activity logs
      responses:
        '200':
          description: Activity logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityLog'
        '403':
          $ref: '#/components/responses/Forbidden'
  /issues/{issueId}/activity:
    get:
      summary: List activity for one issue
      parameters:
        - $ref: '#/components/parameters/IssueId'
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Issue activity
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityLog'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string }
    IssueId:
      in: path
      name: issueId
      required: true
      schema: { type: string }
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Unauthorized:
      description: Missing/invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthErrorResponse' }
    Forbidden:
      description: Role/scope forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ValidationError:
      description: Validation failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    WorkflowStatus:
      type: object
      required: [id, name, order]
      properties:
        id: { type: string }
        name: { type: string }
        order: { type: integer }
    BoardColumn:
      allOf:
        - $ref: '#/components/schemas/WorkflowStatus'
        - type: object
          required: [issues]
          properties:
            issues:
              type: array
              items:
                $ref: '#/components/schemas/Issue'
    UpdateIssueRequest:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        priority: { type: string, enum: [low, medium, high] }
        assigneeId: { type: [string, 'null'] }
        sprintId: { type: [string, 'null'] }
        milestoneId: { type: [string, 'null'] }
        dueAt: { type: [string, 'null'], format: date-time }
        dueDate:
          description: Deprecated compatibility alias of dueAt
          type: [string, 'null']
          format: date-time
    Comment:
      type: object
      required: [id, issueId, body, createdAt]
      properties:
        id: { type: string }
        issueId: { type: string }
        authorId: { type: [string, 'null'] }
        authorName: { type: string }
        authorEmail: { type: [string, 'null'] }
        body: { type: string }
        createdAt: { type: string, format: date-time }
        mentions:
          type: array
          items: { type: string }
    ActivityUserReference:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
    ActivityLog:
      type: object
      required: [id, issueId, action, userReferences, createdAt]
      properties:
        id: { type: string }
        actorId: { type: [string, 'null'] }
        actorName: { type: string }
        actorEmail: { type: [string, 'null'] }
        issueId: { type: string }
        issueNumber: { type: integer }
        issueTitle: { type: string }
        projectId: { type: string }
        projectKey: { type: string }
        projectName: { type: string }
        action: { type: string }
        before: { type: [string, 'null'] }
        after: { type: [string, 'null'] }
        userReferences:
          type: array
          items:
            $ref: '#/components/schemas/ActivityUserReference'
        createdAt: { type: string, format: date-time }
...
    AuthErrorResponse:
      type: object
      properties:
        error:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            status: { type: integer }
...
    Issue:
      type: object
      required: [id, number, projectId, title, description, priority, statusId, dueAt, createdAt, updatedAt]
      properties:
        id: { type: string }
        number: { type: integer }
        projectId: { type: string }
        title: { type: string }
        description: { type: string }
        priority: { type: string, enum: [low, medium, high] }
        statusId: { type: string }
        assigneeId: { type: [string, 'null'] }
        reporterId: { type: [string, 'null'] }
        dueAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CreateIssueRequest:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        priority: { type: string, enum: [low, medium, high] }
        assigneeId: { type: [string, 'null'] }
        reporterId: { type: [string, 'null'] }
        dueAt: { type: [string, 'null'], format: date-time }
        dueDate:
          description: Deprecated compatibility alias of dueAt
          type: [string, 'null']
          format: date-time
    IssueResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Issue'
    IssueListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
        meta:
          type: object
          properties:
            page: { type: integer }
            pageSize: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...